+++ gql_code_gen/lib/src/builders/ast_builder.dart
@@ -3,17 +3,11 @@ import "dart:async";
 import "package:build/build.dart";
 
 import "package:gql_code_gen/src/config.dart";
-import "package:gql_code_gen/src/builders/ast.dart";
-import "package:gql_code_gen/src/reader/reader.dart";
-import "package:gql_code_gen/src/writer/writer.dart";
+import "package:gql_code_gen/src/code_builders/ast.dart";
+import "package:gql_code_gen/src/builders/utils/reader.dart";
+import "package:gql_code_gen/src/builders/utils/writer.dart";
 
-/// Builder factory for AST Builder
-Builder astBuilder(
-  BuilderOptions options,
-) =>
-    _AstBuilder();
-
-class _AstBuilder implements Builder {
+class AstBuilder implements Builder {
   @override
   Map<String, List<String>> get buildExtensions => {
         sourceExtension: [astExtension],
@@ -23,7 +17,9 @@ class _AstBuilder implements Builder {
   FutureOr<void> build(BuildStep buildStep) async {
     final doc = await readDocument(buildStep);
 
-    final library = buildAstLibrary(doc);
+    final library = buildAstLibrary(
+      doc,
+    );
 
     return writeDocument(
       library,
