+++ gql_code_gen/lib/src/builders/input_builder.dart
@@ -1,18 +1,13 @@
 import "dart:async";
 
 import "package:build/build.dart";
-import "package:gql_code_gen/src/builders/input.dart";
+
 import "package:gql_code_gen/src/config.dart";
-import "package:gql_code_gen/src/reader/reader.dart";
-import "package:gql_code_gen/src/writer/writer.dart";
+import "package:gql_code_gen/src/code_builders/input.dart";
+import "package:gql_code_gen/src/builders/utils/reader.dart";
+import "package:gql_code_gen/src/builders/utils/writer.dart";
 
-/// Builder factory for Schema Builder
-Builder inputBuilder(
-  BuilderOptions options,
-) =>
-    _InputBuilder();
-
-class _InputBuilder implements Builder {
+class InputBuilder implements Builder {
   @override
   Map<String, List<String>> get buildExtensions => {
         sourceExtension: [inputExtension],
