+++ gql_code_gen/lib/src/builders/scalar_builder.dart
@@ -3,17 +3,11 @@ import "dart:async";
 import "package:build/build.dart";
 
 import "package:gql_code_gen/src/config.dart";
-import "package:gql_code_gen/src/builders/scalar.dart";
-import "package:gql_code_gen/src/reader/reader.dart";
-import "package:gql_code_gen/src/writer/writer.dart";
+import "package:gql_code_gen/src/code_builders/scalar.dart";
+import "package:gql_code_gen/src/builders/utils/reader.dart";
+import "package:gql_code_gen/src/builders/utils/writer.dart";
 
-/// Builder factory for Schema Builder
-Builder scalarBuilder(
-  BuilderOptions options,
-) =>
-    _ScalarBuilder();
-
-class _ScalarBuilder implements Builder {
+class ScalarBuilder implements Builder {
   @override
   Map<String, List<String>> get buildExtensions => {
         sourceExtension: [scalarExtension],
