+++ gql_code_gen/lib/src/builders/utils/writer.dart
