+++ gql_code_gen/lib/src/code_builders/enum.dart
