+++ gql_code_gen/lib/src/code_builders/scalar.dart
